+++ src/evaluation_pipeline.py
@@ -10,7 +10,7 @@ from pytorch_lightning import Trainer, seed_everything
 
 from src import utils
 
-log = utils.get_logger(__name__)
+log = utils.get_pylogger(__name__)
 
 
 def evaluate(config: DictConfig) -> None:
