+++ src/utils/rich_utils.py
@@ -19,6 +19,7 @@ def print_config_tree(
     cfg: DictConfig,
     print_order: Sequence[str] = (
         "datamodule",
+        "taskmodule",
         "model",
         "callbacks",
         "logger",
